+++ src/idpconfgen/libs/libbuild.py
@@ -6,6 +6,7 @@ from functools import partial
 from itertools import cycle
 
 import numpy as np
+from libfuncpy import flatlist, make_iterable
 from numba import njit
 
 # import idpcpp, imported locally at init_faspr_sidechains
@@ -37,8 +38,12 @@ from idpconfgen.libs.libfilter import (
     regex_search,
     )
 from idpconfgen.libs.libio import read_dictionary_from_disk
-from idpconfgen.libs.libparse import get_mers, translate_seq_to_3l
-from idpconfgen.libs.libtimer import ProgressWatcher, timeme
+from idpconfgen.libs.libparse import (
+    get_mers,
+    translate_seq_to_3l,
+    remove_empty_keys,
+    )
+from idpconfgen.libs.libtimer import ProgressCounter, timeme
 
 
 # See TODO at init_faspr_sidechains
@@ -527,25 +532,57 @@ def read_db_to_slices_given_secondary_structure(database, ss_regexes):
     return primary, seq_angles
 
 
-def prepare_slice_dict(primary, inseq, res_tolerance=None, ncores=1):
-    """."""
+def prepare_slice_dict(
+        primary,
+        input_seq,
+        mers_size=(1, 2, 3, 4, 5),
+        res_tolerance=None,
+        ncores=1,
+        ):
+    """
+    Prepare a dictionary mapping chunks to slices in `primary`.
+
+    Parameters
+    ----------
+    primary : str
+        A concatenated version of all primary sequences in the database.
+        In the form of "QWERY|IPASDF", etc.
+
+    input_seq : str
+        The 1-letter code amino-acid sequence of the conformer to construct.
+
+    mers_size : iterable
+        A iterable of integers denoting the size of the chunks to search
+        for. Defaults from 1 to 5.
+
+    res_tolerance : dict
+        A dictionary mapping residue tolerances, for example:
+        {"A": "AIL"}, noting Ala can be replaced by Ile and Leu in the
+        search (this is a dummy example).
+
+    ncores : int
+        The number of processors to use.
+
+    Return
+    ------
+    dict
+        A dict with the given mapping. First key-leve of the dict
+        is the length of the chunks, hence, integers.
+        The second key level are the residue chunks found in the `primary`.
+        A chunk in input_seq but not in `primary` is removed from the
+        dict.
+    """
     res_tolerance = res_tolerance or {}
+    mers_size = make_iterable(mers_size)
 
     log.info('preparing regex xmers')
-    monomers = get_mers(inseq, 1)
-    dimers = get_mers(inseq, 2)
-    trimers = get_mers(inseq, 3)
-    tetramers = get_mers(inseq, 4)
-    pentamers = get_mers(inseq, 5)
 
+    xmers = (get_mers(input_seq, i) for i in mers_size)
+    xmers_flat = flatlist(xmers)
     slice_dict = defaultdict(dict)
-    mers = it.chain(monomers, dimers, trimers, tetramers, pentamers)
 
-    _l = len(monomers) + len(dimers) + len(trimers) \
-        + len(tetramers) + len(pentamers)
-
-    with ProgressWatcher(_l) as PW:
-        for mer in mers:
+    with ProgressCounter(suffix='Searching for xmers: ') as PW:
+        for mer in xmers_flat:
             lmer = len(mer)
             altered_mer = build_regex_substitutions(mer, res_tolerance)
 
@@ -554,7 +591,7 @@ def prepare_slice_dict(primary, inseq, res_tolerance=None, ncores=1):
             slice_dict[lmer][altered_mer] = \
                 regex_forward_with_overlap(primary, merregex)
 
-            # if no entrey was found
+            # if no slices were found
             if not slice_dict[lmer][altered_mer]:
                 slice_dict[lmer].pop(altered_mer)
 
@@ -571,7 +608,7 @@ def prepare_slice_dict(primary, inseq, res_tolerance=None, ncores=1):
             if not slice_dict[lmer][altered_mer_P]:
                 slice_dict[lmer].pop(altered_mer_P)
             # for _s in slice_dict[lmer][mer]:
-                # assert '|' not in inseq[_s]
+                # assert '|' not in input_seq[_s]
             PW.increment()
 
     return slice_dict