+++ tests/test_libparse.py
@@ -353,3 +353,24 @@ def test_fill_list_bad(in1, fill, size, expected):
     """Test fill list."""
     result = libparse.fill_list(in1, fill, size)
     assert result != expected
+
+
+@pytest.mark.parametrize(
+    'in1,expected',
+    [
+        (['1 2\n', '2 3\n'], {1: 2, 2: 3}),
+        (['1 2'], {1: 2}),
+        ([], {}),
+        ]
+    )
+def test_convert_int_float_lines(in1, expected):
+    """."""
+    result = libparse.convert_int_float_lines_to_dict(in1)
+    assert result == expected
+
+
+def test_remove_empty_keys():
+    """."""
+    d = {'1': 1, 1: '1', 'a': '', 'b': [], 'c': {}}
+    libparse.remove_empty_keys(d)
+    assert list(d.keys()) == ['1', 1]